Cleaning. inline all the things! New RW Lock + guard, some fixing of Logger
Changed logger to be faster by using more inline code and also a static
mapping instead of a "map". Moreover cleaning all headers and deleted
windows lean and mean.

+++ src/tasking/taskqueue.cpp
@@ -44,7 +44,7 @@ namespace jimdb
             m_maxSize = size;
         }
 
-        int TaskQueue::size() const
+        size_t TaskQueue::size() const
         {
             return m_tasks.size();
         }
